Allow adding worker nodes with worker machine manifest
This change adds the functionality to add one or more worker machines to
your cluster by submitting a worker machine manifest.  The
aws-infra-controller running in the cluster watches for changes and
compares the number of desired replicas in the manifest to the number of
worker EC2 instances running in AWS and reconciles by provisioning more
instances when needed.  Note: this does not include the functionality to
delete instances if less replicas are desired.

Signed-off-by: Richard Lander <landerr@vmware.com>